requirements.txt
pillow
huggingface-hub
flask